Change Sensei querylog log level to Debug from Info
Hight QPS service (poncho-web) is causing an issue that it spilling a
log file 1G/0.5 hour.  The content of  logs comes from querylog in
Sensei.
We would like to change the logging level from  Info to Debug for
logging each queries.
Qeurylog is useful information for debugging.  When we need it,  we can
enable it through JMX.